Working on PI controller

+++ node2/drivers/include/PIDriver.h
@@ -1,8 +1,8 @@
 struct PID_data {
   int32_t error_sum;
-  double Kp;
-  double Ki;
+  int8_t Kp;
+  int8_t Ki;
   int16_t position;
 };
-void pid_init(double p_factor, double i_factor, struct PID_data *pid);
+void pid_init(int8_t p_factor, int8_t i_factor, struct PID_data *pid);
 int16_t pid_controller(struct PID_data *pid);
